refactor: move things around to mimic nestjs
Move things around to a NestJS-like resource based structure, based on plural and not singular
