+++ db/config.ts
@@ -1,3 +1,4 @@
+import type { Config } from "drizzle-kit";
 export const dbCredentials = {
     host: process.env.POSTGRES_HOST || "0.0.0.0",
     port: parseInt(process.env.POSTGRES_PORT || '5432'),
@@ -7,3 +8,11 @@ export const dbCredentials = {
 }
 
 export const dbCredentialsString = `postgres://${dbCredentials.user}:${dbCredentials.password}@${dbCredentials.host}:${dbCredentials.port}/${dbCredentials.database}`;
+
+export default {
+    out: "./src/db/migrations",
+    schema: "./src/db/schemas/*.ts",
+    breakpoints: false,
+    driver: "pg",
+    dbCredentials
+} satisfies Config;
