+++ package.json
@@ -9,14 +9,16 @@
     "docs:build": "vitepress build docs",
     "docs:preview": "vitepress preview docs",
     "db:up": "docker-compose up -d",
-    "db:generate": "bun drizzle-kit generate:pg",
-    "db:push": "bun drizzle-kit push:pg",
-    "db:seed": "bun run src/db/seed.ts",
-    "db:dump": "bun drizzle-kit dump:pg",
-    "db:studio": "bun drizzle-kit studio"
+    "db:generate": "bun drizzle-kit generate:pg --config=db/config.ts",
+    "db:migrate": "bun run db/migrations/migrate.ts",
+    "db:push": "bun drizzle-kit push:pg --config=db/config.ts",
+    "db:seed": "bun run db/seed.ts",
+    "db:dump": "bun drizzle-kit dump:pg --config=db/config.ts",
+    "db:studio": "bun drizzle-kit studio --config=db/config.ts"
   },
   "dependencies": {
     "drizzle-orm": "^0.28.6",
+    "drizzle-typebox": "^0.1.1",
     "elysia": "latest",
     "postgres": "^3.3.5"
   },
